Reduce shared sync lint refactor churn
